Changes to timetag.py:
Function writeCsum was modified to call a new function flagOmit if the
csum file is to be constructed from the raw pixel coordinates rather than
the corrected coordinates.  The flagOmit function sets some calibration
switch keywords to OMIT in the primary header of the csum file.  This is
for steps that may have actually been done, but they don't affect the raw
pixel coordinates, so it could be misleading if the keywords in the csum
header are set to complete.  Also call close() in writeCsum.
PR 67675

In function timetagBasicCalibration, get the time column using
cosutil.getColCopy, prior to calling cosutil.timelineTimes and
in the calling sequence for timeline.createTimeline, instead of using
events.field() to get the time column.
Trac #645, PR 67204


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@12130 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 1f5e8defb58112753c0af61c83012ccb6ecefb16