A new dialog box was added to betadrizzle.util to explain the consequences of setting 'updatewcs' to 'yes'.  This dialog box gets used by the new trigger rule added to the .cfgspc file for the 'updatewcs' parameter in 'betadrizzle'.  These changes address Ticket #724.
In addition, the update header code was simplified to not use any distortion model except the CD matrix when applying the fit to the input image headers.  This was tested against ACS/WFC and WFC3/UVIS images, but additional testing still needs to be performed using subarray images to insure that this does not introduce additional problems for those images. 


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@13198 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 108d63378c543e3731089d44929274b7091ef8d2