Removed extraneous 'mode' parameters from 'astrodither' tasks.
Fixed a bug in 'skytopix' so that it will work, and so that input RA,Dec values can be in either floating point values or HMS format with or without ':' (like '19:10:33.7403' or '19 10 33.7403'). This required adding a function 'parse_skypos()' to 'tweakutils' to include logic to recognize floating point inputs in addition to using 'radec_hmsdd' to parse HMS string inputs. 


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@14005 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 22a30fc62aa776c3eac71647040156ef292541e6