Changes to concurrent.py:
Method outFlashSetup was renamed to setUpOutFlash for consistency with
other functions with verb/object names.  This was noticed while checking
out the problem that led to PR 69938.
Trac #806, PR 69938

Changes to timeline.py:

The AIRGLOW_WAVELENGTHS dictionary was modified to specify a tuple of
wavelengths for an airglow line, instead of just a single value.  This was
needed primarily for OI 1304, which is a triplet with a range corresponding
to about 380 pixels for an FUV M grating.  Function findPixelRegion uses
the minimum and maximum wavelengths of an airglow line or lines to find the
range of pixels on the detector.
Trac #805, PR 69937

Changes to timetag.py:

Function getWavecalOffsets was modified; wavecorr was added to the calling
sequence, and a test on wavecorr was added to the test on obstype.
In doDqicorr, a check on minmax_doppler was added before calling
psaWcaBoundary.
Trac #818, PR 69977

Changes to wavecal.py:

Function ttFindWavecalSpectrum returns a tuple of four items (as of calcos
version 2.8.5), but if the input array xi has zero length,
ttFindWavecalSpectrum returned only three:  (0., {}, {}).  Now it returns
four items, and the two dictionaries (xd_shift and xd_locns) are populated
with None and 0 respectively for the current segment or for each stripe.
The dictionaries were populated because concurrent.findShifts expects an
entry for these dictionaries for each segment or stripe.
Function ttFindSpec now checks whether cosutil.centerOfQuadratic returns
None for y_temp.
Trac #806, PR 69938

Changes to x1d.py:

In function makeFltCounts, switches["wavecorr"] was added to the argument
list for getWavecalOffsets.
Trac #818, PR 69977


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@14060 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 09a003c03d1a30436549d16ebe1415a951accdd3