This is an experimental fix to prevent any sort of excessive looping in the stdio capture.  It attacks the problem in two areas:
 1) The special logger that replaces sys.stdout is prevented from having its
    write() method called recursively.  This prevents an infinite loop if the
    sys.stdout logger itself is used as the stream for a StreamHandler.

 2) The LogTeeHandler, which captures writes to stdout and sends a log record
    to the appropriate logger is prevented from teeing a record that was
    already handled by its logger, where part of that handling was to send it
    to a StreamHandler using the sys.stdout logger.

If that sounds confusing, and it should, I can draw a chart.  This is handled
in a thread-safe manner and shouldn't cause any difficulties, but if it does
cause any difficulties better to just remove it than try to fix it; it's
already a hack and can't get much better.  The alternative is to just leave in
the recursion protection on write(); this leads to some duplicate log messages,
but at least doesn't allow any infinite loops.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@14167 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: d6e4f94617628a827340f6ee71c11d9185fc04f3