Relevant Trac tickets and PRs: Trac   PR ----  -----  795  69737  Change the way calcos fixes APERTURE = "RelMvReq".  805  69937  Account for the width of the OI 1304 line in calcos.  829  70033  Calcos should check whether a keyword exists before deleting it.
Changes to cosutil.py

Keyword APERTURE is now included in the list of items that are updated
in the primary header from the `info` dictionary.
Trac #795, PR 69737

Changes to extract.py:

In extract1D, the call to fixApertureKeyword has been commented out, since
it should no longer need to be called.  (see the change to cosutil.py)
Trac #795, PR 69737

Changes to spwcs.py:

Method deleteKeywords in SpWcsCorrtag was modified to check for the
existence of keywords before deleting them.  (This was previously done in
SpWcsImage, but the method in SpWcsCorrtag was missed.)  Parentheses were
removed from the argument list in `del` statements.
Trac #829, PR 70033

Changes to timeline.py:

Function createTimeline was modified.  For ACCUM data the count rate was
being divided by the number of elements in tl_time (which is the TIME
column in the TIMELINE table); this was not correct because the counts had
already been divided by the exposure time to get count rate.  For TIME-TAG
data the count rate was being divided by the exposure time, after getting
just the number of counts within one bin.  Instead, the number of counts
in a bin should be divided by the duration of a bin (dt, nominally but not
necessarily one second).
Trac #805, PR 69937


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@14723 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 5fee7c31f4b9d32d9231132e8db00bb6376fa0a2