Changes to calcos.py:
The changes that had been made to calcos 2.17.3 to disable use of the
gain sag table (GSAGTAB) have been removed.  In particular, gsagtab has
been added to the `temp` dictionary in missingRefFiles, and a check on
its existence has been added just after checking for bpixtab.
Trac #761, PR 69143

Changes to getinfo.py:

Function getRefFileNames was modified to remove the code that set
reffiles["gsagtab"] to "N/A".
Trac #761, PR 69143

Changes to fpavg.py:

In __init__ for OutputX1D, self.output_nelem is initialized to 1, instead
of to 0.  In method computeOutputInfo, if the input spectra are empty,
self.output_nelem is set to the extended array length (16384 for FUV,
1274 for NUV) instead of to 0.  If there are input spectra, self.output_nelem
will only be updated if the computed value for the number of elements is
greater than zero, not just greater than or equal to zero.  The problem was
that self.output_nelem is used to set the array length for x1dsum columns
that are arrays, and zero is not a valid length.
Trac #865, PR 70786


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@15596 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: 90fdacdee6bd2b2f22a22e1c600b9ba3200f2b7f