Changes to wavecal.py:
The previous change to findWavecalShift that assumes there is an FPOFFSET
column in the lamptab was removed.  There are pre-SM4 exposures that need
an older lamptab, one that doesn't have an FPOFFSET column.
Trac #924, PR 71881

Function readFpPixelShift was modified to check for the FPOFFSET column
in the lamptab before trying to read it.
Trac #930, PR 72008


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk@19788 fe389314-cf27-0410-b35b-8c050e845b92

Former-commit-id: cad9bf8b613e3b938977f98d1e1cbbd17c67cdc8